+++ Frends.SFTP.ListFiles/Frends.SFTP.ListFiles.Tests/ErrorTests.cs
@@ -10,15 +10,14 @@ using Frends.SFTP.ListFiles.Enums;
 namespace Frends.SFTP.ListFiles.Tests;
 
 [TestFixture]
-public class ErrorTests
+public class ErrorTests : ListFilesTestBase
 {
     [Test]
     public void ListFiles_TestDirectoryNotExistsThrows()
     {
-        var connection = Helpers.GetSftpConnection();
         var input = new Input
         {
-            Directory = "/listfiles/nonexisting/directory",
+            Directory = "/upload/nonexisting/directory",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -26,7 +25,7 @@ public class ErrorTests
         };
             
 
-        var ex = Assert.Throws<SftpPathNotFoundException>(() => SFTP.ListFiles(input, connection, new CancellationToken()));
+        var ex = Assert.Throws<SftpPathNotFoundException>(() => SFTP.ListFiles(input, _connection, new CancellationToken()));
         Assert.AreEqual($"No such file", ex.Message);
 
     }
@@ -38,7 +37,7 @@ public class ErrorTests
         connection.Port = 51644;
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -57,7 +56,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -78,7 +77,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -99,7 +98,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -122,7 +121,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -143,7 +142,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -163,7 +162,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,
@@ -183,7 +182,7 @@ public class ErrorTests
 
         var input = new Input
         {
-            Directory = "/listfiles",
+            Directory = "/upload",
             FileMask = "*.txt",
             IncludeType = IncludeType.File,
             IncludeSubdirectories = true,