+++ Frends.SFTP.ListFiles/Frends.SFTP.ListFiles.Tests/Lib/Helpers.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Text;
 using System.Security.Cryptography;
 using Renci.SshNet;
 using Renci.SshNet.Common;
+using Renci.SshNet.Security;
 using Frends.SFTP.ListFiles.Definitions;
 using Frends.SFTP.ListFiles.Enums;
 
@@ -22,10 +24,10 @@ internal static class Helpers
         {
             ConnectionTimeout = 60,
             Address = _dockerAddress,
-            Port = 2222,
             Username = _dockerUsername,
-            Authentication = AuthenticationType.UsernamePassword,
             Password = _dockerPassword,
+            Port = 2222,
+            Authentication = AuthenticationType.UsernamePassword,
             ServerFingerPrint = null,
         };
 
@@ -38,42 +40,64 @@ internal static class Helpers
         {
             client.Connect();
             for (var i = 1; i <= 3; i++)
-                client.Create("/listfiles/test" + i + ".txt");
-            client.CreateDirectory("/listfiles/subDir");
+                client.Create("/upload/test" + i + ".txt");
+            client.CreateDirectory("/upload/subDir");
             for (var i = 1; i <= 3; i++)
-                client.Create("/listfiles/subDir/test" + i + ".txt");
+                client.Create("/upload/subDir/test" + i + ".txt");
             client.Disconnect();
         }
     }
 
-    internal static string GetServerFingerprintAsSHA256String()
+    internal static Tuple<byte[], byte[]> GetServerFingerPrintAndHostKey()
     {
-        var fingerprint = "";
+        Tuple<byte[], byte[]> result = null;
         using (var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword))
         {
+            client.ConnectionInfo.HostKeyAlgorithms.Clear();
+            client.ConnectionInfo.HostKeyAlgorithms.Add("ssh-rsa", (data) => { return new KeyHostAlgorithm("ssh-rsa", new RsaKey(), data); });
+
             client.HostKeyReceived += delegate (object sender, HostKeyEventArgs e)
             {
-                // First try with SHA256 typed fingerprint
-                using (SHA256 mySHA256 = SHA256.Create())
-                {
-                    fingerprint = Convert.ToBase64String(mySHA256.ComputeHash(e.HostKey));
-                }
+                result = new Tuple<byte[], byte[]>(e.FingerPrint, e.HostKey);
+                e.CanTrust = true;
             };
+            client.Connect();
+            client.Disconnect();
+        }
+        return result;
+    }
+
+    internal static string ConvertToMD5Hex(byte[] fingerPrint)
+    {
+        return BitConverter.ToString(fingerPrint).Replace("-", ":");
+    }
+
+    internal static string ConvertToSHA256Hash(byte[] hostKey)
+    {
+        var fingerprint = "";
+        using (SHA256 mySHA256 = SHA256.Create())
+        {
+            fingerprint = Convert.ToBase64String(mySHA256.ComputeHash(hostKey));
         }
         return fingerprint;
     }
 
-    internal static string GetServerFingerprintAsMD5String()
+    internal static string ConvertToSHA256Hex(byte[] hostKey)
     {
         var fingerprint = "";
-        using (var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword))
+        using (SHA256 mySHA256 = SHA256.Create())
         {
-            client.HostKeyReceived += delegate (object sender, HostKeyEventArgs e)
-            {
-                fingerprint = BitConverter.ToString(e.FingerPrint).Replace("-", ":");
-            };
+            fingerprint = ToHex(mySHA256.ComputeHash(hostKey));
         }
         return fingerprint;
+    }
+
+    internal static string ToHex(byte[] bytes)
+    {
+        StringBuilder result = new StringBuilder(bytes.Length * 2);
+        for (int i = 0; i < bytes.Length; i++)
+            result.Append(bytes[i].ToString("x2"));
+        return result.ToString();
     }
 
     internal static void DeleteTestFiles()
@@ -81,7 +105,7 @@ internal static class Helpers
         using (var sftp = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword))
         {
             sftp.Connect();
-            sftp.ChangeDirectory("/listfiles");
+            sftp.ChangeDirectory("/upload");
             var files = sftp.ListDirectory(".");
             foreach (var file in files)
             {
@@ -97,7 +121,7 @@ internal static class Helpers
                                 sftp.DeleteFile(f.Name);
                             }
                         }
-                        sftp.ChangeDirectory("/listfiles");
+                        sftp.ChangeDirectory("/upload");
                         sftp.DeleteDirectory(file.FullName);
                     }
                     else
@@ -109,8 +133,6 @@ internal static class Helpers
             sftp.Disconnect();
         }
     }
-
-
 
     internal static SshKeyGenerator.SshKeyGenerator GenerateDummySshKey()
     {